+++ src/ladb_opencutlist/ruby/model/groupdef.rb
@@ -4,13 +4,14 @@ module Ladb::OpenCutList
 
   class GroupDef
 
-    attr_accessor :material_id, :material_name, :material_type, :material_color, :material_grained, :part_count, :std_dimension, :std_available, :std_dimension_stipped_name, :std_width, :std_thickness, :max_number, :total_cutting_length, :total_cutting_area, :total_cutting_volume, :total_final_area, :invalid_final_area_part_count, :show_cutting_dimensions, :show_edges, :edge_decremented
+    attr_accessor :material_id, :material_name, :material_display_name, :material_type, :material_color, :material_grained, :part_count, :std_dimension, :std_available, :std_dimension_stipped_name, :std_width, :std_thickness, :max_number, :total_cutting_length, :total_cutting_area, :total_cutting_volume, :total_final_area, :invalid_final_area_part_count, :show_cutting_dimensions, :show_edges, :edge_decremented
     attr_reader :id, :part_defs
 
     def initialize(id)
       @id = id
       @material_id = ''
       @material_name = ''
+      @material_display_name = ''
       @material_type = MaterialAttributes::TYPE_UNKNOW
       @material_color = nil
       @material_grained = false
@@ -67,6 +68,7 @@ module Ladb::OpenCutList
           :id => @id,
           :material_id => @material_id,
           :material_name => @material_name,
+          :material_display_name => @material_display_name,
           :material_type => @material_type,
           :material_color => @material_color.nil? ? nil : "#%02x%02x%02x" % [ @material_color.red, @material_color.green, @material_color.blue ],
           :material_grained => @material_grained,